feat(wiki): add an introduction to creating jobs
This adds a guide to the wiki that walks the reader through developing a Job with
Open Job Description. The guide explains elements of the Job Template as it introduces
concepts bit by bit, and shows how to use the CLI to verify and test their Jobs on
their workstation.

Signed-off-by: Daniel Neilson <53624638+ddneilson@users.noreply.github.com>